Vectorized 4  carry, Mul and Add

+++ prmers.cpp
@@ -134,7 +134,7 @@ static size_t transformsize(uint32_t exponent) {
 void precalc_for_p(uint32_t p,
                    std::vector<uint64_t>& digit_weight,
                    std::vector<uint64_t>& digit_invweight,
-                   std::vector<int>& digit_width) {
+                   std::vector<uint64_t>& digit_width) {
     size_t n = transformsize(p);
 
     digit_weight.resize(n);
@@ -162,7 +162,7 @@ void precalc_for_p(uint32_t p,
                           << w_val << ") or w+1 (" << w_val + 1 << ") at j = " << j << std::endl;
                 exit(1);
             }
-            digit_width[j - 1] = c;
+            digit_width[j - 1] = static_cast<uint64_t>(c);
             if (j < n) {
                 uint32_t r = (uint32_t)(qj % n);
                 uint64_t nr2r = (r != 0U) ? powModP(nr2, (uint64_t)(n - r)) : 1ULL;
@@ -516,7 +516,7 @@ void executeFusionneNTT_Inverse(cl_command_queue queue,
 }
 
 
-void printVector(const std::vector<int>& vec, const std::string& name) {
+void printVector(const std::vector<uint64_t>& vec, const std::string& name) {
     std::cout << name << " = [ ";
     for (const auto& val : vec) {
         std::cout << val << " ";
@@ -531,7 +531,7 @@ void printVector2(const std::vector<uint64_t>& vec, const std::string& name) {
     std::cout << "]" << std::endl;
 }
 
-void handleFinalCarry(std::vector<uint64_t>& x, const std::vector<int>& digit_width_cpu, size_t n) {
+void handleFinalCarry(std::vector<uint64_t>& x, const std::vector<uint64_t>& digit_width_cpu, size_t n) {
     x[0] += 1;
     uint64_t c = 0;
     
@@ -677,7 +677,7 @@ int main(int argc, char** argv) {
     }
     std::vector<cl_device_id> devices(numDevices);
     clGetDeviceIDs(platform, CL_DEVICE_TYPE_GPU, numDevices, devices.data(), nullptr);
-    if(device_id < 0 || device_id >= static_cast<int>(numDevices)) {
+    if (device_id < 0 || static_cast<uint64_t>(device_id) >= numDevices){
         std::cerr << "Invalid device id specified (" << device_id << "). Using device 0 instead." << std::endl;
         device_id = 0;
     }
@@ -718,7 +718,7 @@ int main(int argc, char** argv) {
 
     // Precompute parameters
     std::vector<uint64_t> digit_weight_cpu, digit_invweight_cpu;
-    std::vector<int> digit_width_cpu;
+    std::vector<uint64_t> digit_width_cpu;
     precalc_for_p(p, digit_weight_cpu, digit_invweight_cpu, digit_width_cpu);
     if(debug){
         printVector2(digit_weight_cpu, "digit_weight_cpu");
@@ -864,7 +864,7 @@ int main(int argc, char** argv) {
 
     cl_mem buf_digit_width = createBuffer(context,
         CL_MEM_READ_ONLY | CL_MEM_COPY_HOST_PTR,
-        digit_width_cpu.size() * sizeof(int),
+        digit_width_cpu.size() * sizeof(uint64_t),
         (void*)digit_width_cpu.data(), "buf_digit_width");
 
     cl_mem buf_twiddles = createBuffer(context,
